+++ internal/cli/zones_distribution.go
@@ -25,6 +25,8 @@ func (d *distributionOutput) TableRows() [][]string {
 
 func (d *distributionOutput) JSONData() any { return d }
 
+func (d *distributionOutput) TemplateData() any { return d.Data }
+
 // zoneFileOutput adapts ZoneFile for output.
 type zoneFileOutput struct {
 	Data *dnsimple.ZoneFile `json:"data"`
@@ -39,6 +41,8 @@ func (z *zoneFileOutput) TableRows() [][]string {
 }
 
 func (z *zoneFileOutput) JSONData() any { return z }
+
+func (z *zoneFileOutput) TemplateData() any { return z.Data }
 
 func newZonesFileCmd(f *cmdutil.Factory) *cobra.Command {
 	return &cobra.Command{